requirements.txt
# Django
asgiref==3.8.1
Django==5.2.1
sqlparse==0.5.3
tzdata==2025.2

# Driver PostgreSQL (psycopg)
psycopg==3.2.9
psycopg-binary==3.2.9

# Manejo de imágenes (pillow)
pillow==11.2.1

# Manejo de países (django-countries)
django-countries==7.6.1
typing_extensions==4.14.0

# Manejo de teléfonos (django-phonenumber-field)
django-phonenumber-field==8.1.0
phonenumbers==9.0.6

# Limpieza de huérfanos (django-cleanup)
django-cleanup==9.0.0

# Interfaz de administración responsive (django-jet-reboot)
django-jet-reboot==1.3.10

# Meteorología (yr-weather)
attrs==25.3.0
cattrs==25.1.1
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
platformdirs==4.3.8
pytz==2025.2
requests==2.32.4
requests-cache==1.2.1
url-normalize==2.2.1
urllib3==2.5.0
xmltodict==0.14.2
yr-weather==0.4.0

# Geodistancia (haversine)
haversine==2.9.0

# Cobertura de código (coverage)
coverage==7.10.4

# Métricas de código (radon)
colorama==0.4.6
mando==0.7.1
radon==6.0.1
six==1.17.0

# Ratings (glicko-2)
glicko2==2.1.0

# Mapas (django-leaflet)
django-leaflet==0.32.0

# Autenticación avanzada (django-allauth)
django-allauth==65.11.0

# Códigos QR (django-qr-code)
annotated-types==0.7.0
django-qr-code==4.2.0
pydantic==2.11.7
pydantic_core==2.33.2
segno==1.6.6
typing-inspection==0.4.1

# Encuestas (django-form-surveys)
django-form-surveys==2.5.0
django-tinymce==4.1.0

# Permisos a nivel de objeto (django-guardian)
django-guardian==3.0.3

# Exportación de modelos a PlantUML (django-model2puml)
django-model2puml==0.6.1
